+++ hello.js
@@ -1,0 +1,5 @@
+export class Hello {
+    static sayHello() {
+        console.log('Hello World');
+    }
+}