+++ tests/specs/references/class/staticMemberWithArguments.d.ts
@@ -1,0 +1,5 @@
+export class Hello {
+    static SayHelloTo (name: string): void;
+    static SayHelloTo2Persons (name1: string, name2: string): void;
+    static SayHelloTo3Persons (name1: string, name2: string, name3: string): void;
+}